dagster-informatics-companies/pyproject.toml
[project]
name = "dagster_informatics_companies"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.13.0",
    "dagster-dbt>=0.29.0",
    "dbt-postgres>=1.10.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "requests>=2.33.1",
    "xlrd>=2.0.2",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dagster_informatics_companies"
registry_modules = [
    "dagster_informatics_companies.components.*",
]


